add resource check in rds mysql sg-destroy script

+++ infra/scripts/dbs/rds/sg-destroy.sh
@@ -2,4 +2,9 @@
 GIT_TLD=`git rev-parse --show-toplevel`
 source ${GIT_TLD}/.env
 RDS_VPC_SECURITY_GROUP_ID=$(aws ec2 describe-security-groups --region ${AWS_REGION}  --filters Name=group-name,Values=RobotShopRDSSecurityGroup --query 'SecurityGroups[*].GroupId' --output text --no-cli-pager)
-aws ec2 delete-security-group --group-id $RDS_VPC_SECURITY_GROUP_ID --region ${AWS_REGION} --no-cli-pager
+
+if [ -n "$RDS_VPC_SECURITY_GROUP_ID" ]; then
+    aws ec2 delete-security-group --group-id $RDS_VPC_SECURITY_GROUP_ID --region ${AWS_REGION} --no-cli-pager
+else
+    echo "MySQL security group does not exist."
+fi
